+++ src/cls/CMakeLists.txt
@@ -74,6 +74,10 @@ if (WITH_RADOSGW)
   set(cls_otp_srcs otp/cls_otp.cc)
   add_library(cls_otp SHARED ${cls_otp_srcs})
   target_link_libraries(cls_otp OATH::OATH)
+  target_include_directories(cls_otp
+	  PUBLIC "${CMAKE_SOURCE_DIR}/src/rgw/driver/rados"
+	  PUBLIC "${CMAKE_SOURCE_DIR}/src/rgw"
+	  PUBLIC "${CMAKE_SOURCE_DIR}/src/spawn/include")
   set_target_properties(cls_otp PROPERTIES
     VERSION "1.0.0"
     SOVERSION "1"
@@ -198,6 +202,10 @@ if (WITH_RADOSGW)
     ${CMAKE_SOURCE_DIR}/src/common/ceph_json.cc)
   add_library(cls_rgw SHARED ${cls_rgw_srcs})
   target_link_libraries(cls_rgw json_spirit)
+  target_include_directories(cls_rgw
+	  PUBLIC "${CMAKE_SOURCE_DIR}/src/rgw/driver/rados"
+	  PUBLIC "${CMAKE_SOURCE_DIR}/src/rgw"
+	  PUBLIC "${CMAKE_SOURCE_DIR}/src/spawn/include")
   set_target_properties(cls_rgw PROPERTIES
     VERSION "1.0.0"
     SOVERSION "1"
@@ -210,6 +218,10 @@ if (WITH_RADOSGW)
     rgw/cls_rgw_types.cc
     rgw/cls_rgw_ops.cc)
   add_library(cls_rgw_client STATIC ${cls_rgw_client_srcs})
+  target_include_directories(cls_rgw_client
+	  PUBLIC "${CMAKE_SOURCE_DIR}/src/rgw/driver/rados"
+	  PUBLIC "${CMAKE_SOURCE_DIR}/src/rgw"
+	  PUBLIC "${CMAKE_SOURCE_DIR}/src/spawn/include")
 
 endif (WITH_RADOSGW)
 
@@ -299,6 +311,10 @@ if (WITH_RADOSGW)
     queue/cls_queue_src.cc
     ${CMAKE_SOURCE_DIR}/src/common/ceph_json.cc)
   add_library(cls_rgw_gc SHARED ${cls_rgw_gc_srcs})
+  target_include_directories(cls_rgw_gc
+	  PUBLIC "${CMAKE_SOURCE_DIR}/src/rgw/driver/rados"
+	  PUBLIC "${CMAKE_SOURCE_DIR}/src/rgw"
+	  PUBLIC "${CMAKE_SOURCE_DIR}/src/spawn/include")
   set_target_properties(cls_rgw_gc PROPERTIES
     VERSION "1.0.0"
     SOVERSION "1"
@@ -309,6 +325,10 @@ if (WITH_RADOSGW)
   set(cls_rgw_gc_client_srcs
     rgw_gc/cls_rgw_gc_client.cc)
   add_library(cls_rgw_gc_client STATIC ${cls_rgw_gc_client_srcs})
+  target_include_directories(cls_rgw_gc_client
+	  PUBLIC "${CMAKE_SOURCE_DIR}/src/rgw/driver/rados"
+	  PUBLIC "${CMAKE_SOURCE_DIR}/src/rgw"
+	  PUBLIC "${CMAKE_SOURCE_DIR}/src/spawn/include")
 endif (WITH_RADOSGW)
