+++ src/rgw/rgw_multi.h
@@ -6,6 +6,7 @@
 
 #include <map>
 #include "rgw_xml.h"
+#include "rgw_obj_types.h"
 #include "rgw_obj_manifest.h"
 #include "rgw_compression_types.h"
 #include "common/dout.h"
@@ -13,51 +14,6 @@
 
 #define MULTIPART_UPLOAD_ID_PREFIX_LEGACY "2/"
 #define MULTIPART_UPLOAD_ID_PREFIX "2~" // must contain a unique char that may not come up in gen_rand_alpha()
-
-class RGWMPObj;
-
-struct RGWUploadPartInfo {
-  uint32_t num;
-  uint64_t size;
-  uint64_t accounted_size{0};
-  std::string etag;
-  ceph::real_time modified;
-  RGWObjManifest manifest;
-  RGWCompressionInfo cs_info;
-
-  RGWUploadPartInfo() : num(0), size(0) {}
-
-  void encode(bufferlist& bl) const {
-    ENCODE_START(4, 2, bl);
-    encode(num, bl);
-    encode(size, bl);
-    encode(etag, bl);
-    encode(modified, bl);
-    encode(manifest, bl);
-    encode(cs_info, bl);
-    encode(accounted_size, bl);
-    ENCODE_FINISH(bl);
-  }
-  void decode(bufferlist::const_iterator& bl) {
-    DECODE_START_LEGACY_COMPAT_LEN(4, 2, 2, bl);
-    decode(num, bl);
-    decode(size, bl);
-    decode(etag, bl);
-    decode(modified, bl);
-    if (struct_v >= 3)
-      decode(manifest, bl);
-    if (struct_v >= 4) {
-      decode(cs_info, bl);
-      decode(accounted_size, bl);
-    } else {
-      accounted_size = size;
-    }
-    DECODE_FINISH(bl);
-  }
-  void dump(Formatter *f) const;
-  static void generate_test_instances(std::list<RGWUploadPartInfo*>& o);
-};
-WRITE_CLASS_ENCODER(RGWUploadPartInfo)
 
 class RGWMultiCompleteUpload : public XMLObj
 {
